Add test coverage and rename file for viewHistory
