+++ packages/rum-core/src/domain/contexts/internalContext.ts
@@ -1,7 +1,7 @@
 import type { RelativeTime } from '@datadog/browser-core'
 import type { ActionContexts } from '../action/actionCollection'
 import type { RumSessionManager } from '../rumSessionManager'
-import type { ViewHistory } from './viewHistoryEntries'
+import type { ViewHistory } from './viewHistory'
 import type { UrlContexts } from './urlContexts'
 
 export interface InternalContext {
@@ -25,13 +25,13 @@ export interface InternalContext {
 export function startInternalContext(
   applicationId: string,
   sessionManager: RumSessionManager,
-  viewHistoryEntries: ViewHistory,
+  viewHistory: ViewHistory,
   actionContexts: ActionContexts,
   urlContexts: UrlContexts
 ) {
   return {
     get: (startTime?: number): InternalContext | undefined => {
-      const viewContext = viewHistoryEntries.findView(startTime as RelativeTime)
+      const viewContext = viewHistory.findView(startTime as RelativeTime)
       const urlContext = urlContexts.findUrl(startTime as RelativeTime)
       const session = sessionManager.findTrackedSession(startTime as RelativeTime)
       if (session && viewContext && urlContext) {
