+++ packages/rum-core/src/domain/contexts/viewHistory.ts
