+++ packages/rum-core/src/domain/getSessionReplayUrl.ts
@@ -1,6 +1,6 @@
 import { INTAKE_SITE_STAGING, INTAKE_SITE_US1, INTAKE_SITE_EU1 } from '@datadog/browser-core'
 import type { RumConfiguration } from './configuration'
-import type { ViewHistoryEntry } from './contexts/viewHistoryEntries'
+import type { ViewHistoryEntry } from './contexts/viewHistory'
 import type { RumSession } from './rumSessionManager'
 
 export function getSessionReplayUrl(
