+++ packages/rum-core/test/testSetupBuilder.ts
@@ -15,7 +15,7 @@ import { validateAndBuildRumConfiguration } from '../src/domain/configuration'
 import type { FeatureFlagContexts } from '../src/domain/contexts/featureFlagContext'
 import type { PageStateHistory } from '../src/domain/contexts/pageStateHistory'
 import type { UrlContexts } from '../src/domain/contexts/urlContexts'
-import type { ViewHistory } from '../src/domain/contexts/viewHistoryEntries'
+import type { ViewHistory } from '../src/domain/contexts/viewHistory'
 import type { RawRumEventCollectedData } from '../src/domain/lifeCycle'
 import { LifeCycle } from '../src/domain/lifeCycle'
 import type { ActionContexts } from '../src/domain/action/actionCollection'
